+++ build-node.sh
@@ -13,18 +13,19 @@ import { Buffer } from 'buffer';
 globalThis.Buffer = Buffer;
 
 export * as Wasm from './wasm';
+export { Kiwi } from './kiwi';
+export { Rpc } from './rpc/client';
+export { KRC20 } from './krc20';
+export { Utils, Enum } from './utils';
+
 export * from "./address/mnemonic";
 export * from "./address/wallet";
-export * from "./rpc/client";
 export * from "./api/kasplexApi";
 export * from "./api/kaspaApi";
 export * from "./KaspaTransaction";
 export * from "./script/script";
-export * from "./krc20";
-export * from './kiwi';
 export * from './init';
-export * from "./utils/index";
-export * as Modules from "./types/index";
+export * as Modules from "./types";
 export * as Tx from "./tx";
 EOF